requirements/dev.txt
-r base.txt

ipython==7.4.0
Werkzeug==0.15.2
django-extensions==2.1.6